feat: 가게정보상세 리팩토링
