+++ src/app/(main)/owner/page.tsx
@@ -1,22 +1,22 @@
-"use client";
+'use client';
 
-import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
+import { useEffect, useState } from 'react';
+import { useRouter } from 'next/navigation';
 
-import { EmptySection } from "@/components/common/EmptySection";
-import { getUser } from "@/api/users";
-import { useAuthStore } from "@/stores/auth";
+import { EmptySection } from '@/components/common/EmptySection';
+import { getUser } from '@/api/users';
+import { useAuthStore } from '@/stores/auth';
 
-import type { UserDetailResponse } from "@/types/user";
-import type { Shop } from "@/types/shop";
+import type { UserDetailResponse } from '@/types/user';
+import type { Shop } from '@/types/shop';
 
-type ViewState = "idle" | "loading" | "noShop" | "error";
+type ViewState = 'idle' | 'loading' | 'noShop' | 'error';
 
 function isUserDetailResponse(data: unknown): data is UserDetailResponse {
-  if (!data || typeof data !== "object") return false;
+  if (!data || typeof data !== 'object') return false;
 
   const obj = data as { item?: unknown };
-  if (!obj.item || typeof obj.item !== "object") return false;
+  if (!obj.item || typeof obj.item !== 'object') return false;
 
   const item = obj.item as {
     id?: unknown;
@@ -25,7 +25,7 @@ function isUserDetailResponse(data: unknown): data is UserDetailResponse {
     shop?: { item?: unknown } | null;
   };
 
-  const isString = (v: unknown) => typeof v === "string";
+  const isString = (v: unknown) => typeof v === 'string';
 
   return isString(item.id) && isString(item.email);
 }
@@ -34,9 +34,9 @@ export default function OwnerPage() {
   const router = useRouter();
 
   const user = useAuthStore((state) => state.user);
-  const userId = user?.id ?? "";
+  const userId = user?.id ?? '';
 
-  const [viewState, setViewState] = useState<ViewState>("idle");
+  const [viewState, setViewState] = useState<ViewState>('idle');
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
 
   useEffect(() => {
@@ -44,14 +44,14 @@ export default function OwnerPage() {
 
     async function checkMyShop() {
       try {
-        setViewState("loading");
+        setViewState('loading');
         setErrorMessage(null);
 
         const userRes = await getUser(userId);
         const userData = userRes.data;
 
         if (!isUserDetailResponse(userData)) {
-          throw new Error("예상치 못한 사용자 응답 형식입니다.");
+          throw new Error('예상치 못한 사용자 응답 형식입니다.');
         }
 
         const me = userData.item;
@@ -64,12 +64,12 @@ export default function OwnerPage() {
         }
 
         // 가게가 없으면 → 가게 등록 화면
-        setViewState("noShop");
+        setViewState('noShop');
       } catch (error: unknown) {
         console.error(error);
-        let msg = "내 가게 정보를 확인하는 중 오류가 발생했습니다.";
+        let msg = '내 가게 정보를 확인하는 중 오류가 발생했습니다.';
 
-        if (error && typeof error === "object" && "response" in error) {
+        if (error && typeof error === 'object' && 'response' in error) {
           const e = error as { response?: { data?: { message?: string } } };
           msg = e.response?.data?.message ?? msg;
         } else if (error instanceof Error && error.message) {
@@ -77,7 +77,7 @@ export default function OwnerPage() {
         }
 
         setErrorMessage(msg);
-        setViewState("error");
+        setViewState('error');
       }
     }
 
@@ -99,27 +99,25 @@ export default function OwnerPage() {
   }
 
   // 2) 로딩 중
-  if (viewState === "idle" || viewState === "loading") {
+  if (viewState === 'idle' || viewState === 'loading') {
     return (
       <main className="w-full bg-white flex flex-col items-center">
         <section className="py-15 w-full max-w-[964px] flex flex-col">
           <span className="tj-h1 text-gray-black">내 가게</span>
-          <p className="mt-4 tj-body1 text-gray-50">
-            내 가게 정보를 확인하는 중입니다...
-          </p>
+          <p className="mt-4 tj-body1 text-gray-50">내 가게 정보를 확인하는 중입니다...</p>
         </section>
       </main>
     );
   }
 
   // 3) 에러
-  if (viewState === "error") {
+  if (viewState === 'error') {
     return (
       <main className="w-full bg-white flex flex-col items-center">
         <section className="py-15 w-full max-w-[964px] flex flex-col">
           <span className="tj-h1 text-gray-black">내 가게</span>
           <p className="mt-4 tj-body1 text-red-500">
-            {errorMessage ?? "내 가게 정보를 확인하는 중 문제가 발생했습니다."}
+            {errorMessage ?? '내 가게 정보를 확인하는 중 문제가 발생했습니다.'}
           </p>
         </section>
       </main>
@@ -127,7 +125,7 @@ export default function OwnerPage() {
   }
 
   // 4) 가게 없음 → 가게 등록 유도
-  if (viewState === "noShop") {
+  if (viewState === 'noShop') {
     return (
       <main className="w-full bg-white flex flex-col items-center">
         <section className="py-15 w-full max-w-[964px] flex flex-col">
@@ -135,7 +133,7 @@ export default function OwnerPage() {
             title="내 가게"
             description="내 가게를 소개하고 공고도 등록해 보세요."
             buttonLabel="가게 등록하기"
-            href="/owner/shops/[shopId]/new"
+            href="/owner/shops/new"
           />
         </section>
       </main>
